unreported bug fix: ERROR	 KNIME-Worker-0 WorkflowEditor	 Failed to delete auto-saved copy of workflow (folder "/tmp/knime_temp_space_wiswedel_82384/Scorer82386/.Scorer.autoSave"): File does not exist: /tmp/knime_temp_space_wiswedel_82384/Scorer82386/.Scorer.autoSave/MISSING Table to Spark (#186)/settings.xml Exception in thread "Delete temporary copy of knime://Testflows/BigDataTests/trunk/FlakyTestflows/SparkExecutor/spark_all/Scorer" java.lang.IllegalArgumentException: /tmp/knime_temp_space_wiswedel_82384/Scorer82386/.Scorer.autoSave/MISSING Spark Scorer (#221) does not exist 	at org.apache.commons.io.FileUtils.cleanDirectory(FileUtils.java:1637) 	at org.apache.commons.io.FileUtils.deleteDirectory(FileUtils.java:1535) 	at org.apache.commons.io.FileUtils.forceDelete(FileUtils.java:2270) 	at org.apache.commons.io.FileUtils.cleanDirectory(FileUtils.java:1653) 	at org.apache.commons.io.FileUtils.deleteDirectory(FileUtils.java:1535) 	at org.apache.commons.io.FileUtils.forceDelete(FileUtils.java:2270) 	at org.apache.commons.io.FileUtils.cleanDirectory(FileUtils.java:1653) 	at org.apache.commons.io.FileUtils.deleteDirectory(FileUtils.java:1535) 	at org.knime.workbench.editor2.WorkflowEditor.lambda$0(WorkflowEditor.java:535) 	at java.lang.Thread.run(Thread.java:745)

+++ org.knime.workbench.editor/src/org/knime/workbench/editor2/WorkflowEditor.java
@@ -531,7 +531,7 @@ public class WorkflowEditor extends GraphicalEditor implements
                         try {
                             m_workflowCanBeDeleted.acquire();
                             File d = flowLoc.toLocalFile();
-                            if (d != null) {
+                            if (d != null && d.exists()) {
                                 FileUtils.deleteDirectory(d.getParentFile());
                             }
                         } catch (CoreException | IOException | InterruptedException e) {
